mousescroll z

+++ Data/Avatar.cs
@@ -14,9 +14,9 @@ namespace WorldOfAvatars.Data
         public string? AvatarSvg { get; set; }
         public bool Mine { get; set; } = false;
 
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
+        public float Z { get; set; }
 
         // public Avatar(string userName, string avatarSvg, bool mine)
         // {
